+++ src/mc-efc-private.h
@@ -32,6 +32,9 @@ typedef enum _supported_query_type_flags {
     SUPPORTS_SUBSTRING_PREVIEW_QUERIES = 1 << 3,
     SUPPORTS_SUFFIX_QUERIES = 1 << 4,
     SUPPORTS_PREFIX_QUERIES = 1 << 5,
+    // prefixPreview and suffixPreview are dropped. Setting this results in an error.
+    SUPPORTS_SUFFIX_PREVIEW_DEPRECATED_QUERIES = 1 << 6,
+    SUPPORTS_PREFIX_PREVIEW_DEPRECATED_QUERIES = 1 << 7,
 } supported_query_type_flags;
 
 typedef struct _mc_EncryptedField_t {